+++ CHANGELOG.md
@@ -40,6 +40,7 @@ This project adheres to [Semantic Versioning](https://semver.org/).
 
 ### Fixed
 
+- Disable caching for the fake frontend (#2322)
 - Avoid undefined array access in `SystemEmailFromBuilder` (#2310)
 - Provide the request to `newCObj()` for TYPO3 V13 (#2281)
 - Avoid the deprecated `TypoScriptFrontendController::getLanguage()` (#2278)